+++ CHANGELOG.md
@@ -13,8 +13,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
--   Drop Laravel > 8 support.
+-   Drop Laravel >= 7 support.
 -   Remove `header` versioning method.
+-   Remove `base_namespace` and `base_prefix` config.
+
+    -   `Route::lapiv()` received an overhaul and it has been decided that these configurations shouldn't be handled by `Route::lapiv()`
+    -   If you want to retain the old behaviour, you'll need to modify `Route::lapiv()` like so:
+
+        ```php
+        Route::namespace('\App\Http\Controllers\Api')
+            ->group(function () {
+                Route::lapiv(function (){
+                    Route::get('foo', 'FooGatewayController@index');
+                    Route::get('foo/{foo}', 'FooGatewayController@show');
+                })
+            });
+        ```
 
 ## v1.0.1 (2021-01-08)